fix encounter stats broken by talent rework
