+++ src/CHANGELOG.tsx
@@ -67,6 +67,7 @@ import SpellLink from 'interface/SpellLink';
 import { TALENTS_SHAMAN } from 'common/TALENTS';
 
 export default [
+  change(date(2022, 11, 29), 'Fix encounter stats broken by the talent rework.', ToppleTheNun),
   change(date(2022, 11, 28), 'Add chevron to cooldown expandables.', ToppleTheNun),
   change(date(2022, 11, 26), 'Add potions for Dragonflight.', ToppleTheNun),
   change(date(2022, 11, 26), 'Add gems for Dragonflight.', ToppleTheNun),
